Add Delegate method and variable to CannonManager
CannonManager now contains a delgate method that
returns void and accepts an int parameter.
An instance of the delgate, OnGameStateChange,
has been added. SetGameState is subscribed to the
delegate.

A new method in CannonBallLoadingManager, CheckGameState
is subscribed to CannonManager.OnGameStateChange as well.
CheckGameState will call the appropriate functions depending
on the value of CannonManager._gameState.