Modify CannonManager and CannonBallLoadingManager
CannonManager.SetGameState now checks for a change in
_gameState value before invoking the OnGameStateChange
delagate variable. SetUp method removed. SetGameState
and GetGameState methods are now public static methods.

CannonBallLoadingManager.CheckGameState method is
subscribed to CannonManager.OnGameStateChange.
There is a switch case that checks for states 1, 2, and 3.

State 1, the "replenishment" phase, calls SetUpLoadingPhase,
and then calls CannonManager.SetGameState(2).

State 2 is "loading" phase, and sets the
SelectableCannonBallZones' isTargetable to true.

State 3 is the "aiming" phase, and sets the
SelectableCannonBallZones' isTargetable to false.