Modify PhaseButtonManager and CannonProperties
PhaseButtonManager now has several methods that
sets the phase button's sprites depending on
the game state. PhaseButtonManager subscribes
CheckGameState to CannonManager.SetGameState.

CannonProperties has a delegate, OnCheck, that
PhaseButtonManager is subscribed to. When the
cannon has at least one ammo in it or has no
ammo in it, it will call the
PhaseButtonManager.SetIntermediateAimingStates.

GameProperties has a dictionary containing all
the phase buttons' sprites. These are accessed
through GetPhaseSprite.