Refactor PhaseButtonManager and GameProperties
As button enabled states will also be modified
based on game state, for loops have been added
to go through the arrays of sprites and
colliders.

_buttonStateCombinArray contains all possible
combinations of expected button sprite
configurations. A button's collider is only
when enabled when the sprite is set to
index state 1 or 2.

GameProperties._phaseSpritesDict uses integers
instead of strings for the key.