Modify CannonProperties and TileProperties
TileProperties has a new method, GetTileCoordinates.
It returns the tile coordinates of the tile.

CannonProperties has a new method, DelistTarget.
It checks the tile coordinates of each tile in
_selectedTargets. If the clicked tile and a tile
in _selectedTargets have matching coordinates,
then the tile is removed from the list.

TileProperties.CheckGameState case 4 now calls
CannonProperties.DelistTarget when _isDesignated
is false.