Modify CannonProperties and TileProperties
CannonProperties.DesignateTarget now checks whether
the number of selected targets is less than the
number of loaded cannonballs. If this is true, then
the tile is added to _selectedTargets and returns true.
If false, then DesignateTargetreturns false.

CannonProperties.DelistTarget returns true if a tile
match is found, and returns false if no tile is matched.

TileProperties.CheckGameState case 4 now has a bool
value, isTaskComplete, that is used to check whether to change
_isDesignated. isTaskComplete must be true for _isDesignated to
be changed.